Migrite mahara_oauth class to its own file and namespace.
Resolves #28
